Add `templates.default.includeMethodsIndex` config option
When this option is set to `true`, a new section is added to the rendered HTML
pages.

This section, entitled "_Methods_", provides a list of methods included in the
file with internal links to the related section.

The section is inserted on the side of the first displayed method, or the
constructor.

_Private methods_ are displayed with an _italic_ font.

Adds as well a little tweak that highlights slightly the text displayed into a
"`code`" HTML tag.